Change React Intl Components to be self-closing
This modifies all of the React Intl Components to be self-closing
instead of relying on the nested child to be the value of interest. All
of the components (beside Message and HTMLMessage) have a `value` prop:

```
<FromattedNumber value={1000} />
```

The Message and HTMLMessage components use a `message` prop instead of
`value`. Both of these components now support React Elements as prop
values! This makes formatting message _much_ more powerful, and for the
Message component, React will be able to use virtual DOM diffing. For
the HTMLMessage component there will be a perf hit since the entire
message must end up as a big HTML string that's then injected into the
DOM via `dangerouslySetInnerHTML`.

```
var msg = "Posted {ago}, {likes, plural, one{# like} other{# likes}}";

<FormattedMessage message={msg}
   ago={<FormattedRelative value={post.date} />}
   likes={post.likes} />
```

Details of the discussion on using nested children vs. self-closing
elements, and supporting React Elements in messages is in #65.

Fixes #26