Properly resolve parent locales
This is a major, additive, non-breaking change that properly resolves
and exposes parent locale information.

The CLDR uses an [algorithm to resolve locale data][CLDR resolve], but
there are [exceptions][CLDR parent locales]. This change leverages the
data for the exceptional cases to now properly resolve a locale's
parent locale.

Locale entries with a parent will now contain a `parentLocale`
property. This can be used by code consuming the data this package
outputs to accurately traverse the locale hierarchy. This is very
useful since this package will de-duplicate data as much as possible.

Also, this change now guarantees that all locales listed in the
`locales` option will have entries in the output data.

[CLDR resolve]: http://cldr.unicode.org/development/development-process/design-proposals/resolution-of-cldr-files
[CLDR parent locales]: http://cldr.unicode.org/development/development-process/design-proposals/parent-locales