Refactor build and package layout
This is a major refactor of React Intl's build and npm package
layout.

Locale data is now package-relative and can more easily be loaded:
```js
import enLocaleData from 'react-intl/locale-data/en';
```

There is now only one copy of these locale data files in the package
and they are UMD modules which export the locale data for a language.

Rollup is now being used to build a _single-file_ React Intl module
for three different targets:

- `ReactIntl` global: Browser build with bundled dependencies
- CommonJS (`main`): For Node/Browserify/Webpack with `require()`-d
dependencies
- ES6 (`jsnext:main`): For consumption via `import` with `import`-d
dependencies

All of the builds are compiled with Babel and bundled into a single
module file. Rollup is used to fold all of the internal `src/`
modules into one. The difference between `lib/index.js` and
`lib/index.es.js` is the latter has everything compiled expect the
`import`/`export` statements for React Intl as a whole package.

The `.babelrc` files have been moved to closer to the source code
which they apply. This allows for test or example code to leverage
experimental ECMAScript syntax without leaking it into the main src.

These changes should result is a smaller and faster package as all of
React Intl's code is folded into a single module scope, and the Babel
helpers are only included once (instead of per module file). As well
as fixing support for IE8.

Fixes #203
Fixes #204
Fixes #275
Fixes #277