Use specific thresholds for relative formatting
The `intl-relativeformat` package has a set of thresholds to round-up
to the next significant unit that make sense for static formatting.
But these thresholds don't make sense when `<FormattedRelative>`
automatically "ticks" the relative times.

This change adjusts the thresholds to not round-up, and does so in a
way that shouldn't effect other uses of `intl-relativeformat` because
the thresholds are only temporarily updated, then put back after
formatting.