feat(@formatjs/cli): revamp underlying extraction We replaced the underlying AST extractor from babel to typescript. This makes the extractor a lot more future-proof due to babel having to constantly include new plugin/syntax while TS just works out of the box. Future new syntax would just be TS upgrade.
BREAKING CHANGE: Remove `--messages-dir` option. This was primarily used
to eagerly write out output in the babel plugin since we don't know when
the execution will be done. This is not the case with the CLI.
`--out-file` should be used instead.

BREAKING CHANGE: Remove `--module-source-name` option. This is not used.